+++ src/main/java/com/atlauncher/data/modrinth/ModrinthProjectType.java
@@ -27,5 +27,8 @@ public enum ModrinthProjectType {
     MODPACK,
 
     @SerializedName("resourcepack")
-    RESOURCEPACK
+    RESOURCEPACK,
+
+    @SerializedName("shader")
+    SHADER,
 }