+++ CHANGELOG.md
@@ -9,5 +9,6 @@ This changelog only contains the changes that are unreleased. For changes for in
 
 ### Fixes
 - Issue with resource packs not showing files when adding to an instance with a loader [#700]
+- Resource packs from Modrinth not installing correctly [#701]
 
 ### Misc