+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 This changelog only contains the changes that are unreleased. For changes for individual releases, please visit the
 [releases](https://github.com/ATLauncher/ATLauncher/releases) page on GitHub.
 
-## 3.4.23.2
+## 3.4.24.0
 
 ### New Features
+- Add in support for shaders from Modrinth
 
 ### Fixes
 - Issue with resource packs not showing files when adding to an instance with a loader [#700]